requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
fake-useragent==0.1.11
idna==2.9
lxml==4.5.1
PySocks==1.7.1
PyYAML==5.3.1
requests==2.23.0
selenium==3.141.0
soupsieve==2.0.1
stem==1.8.0
tqdm==4.46.1
urllib3==1.25.9
